divide Dto request, response on projects
